requirements.txt
django>=4.2
black>=23.1
opencv-python>=4.5
openai>=0.28
requests>=2.31
django-celery-results>=2.5
python-dotenv>=1.0
Pillow>=10.0
python-decouple
